+++ tests/testthat/test_pipelines/test_pipeline_monthly_monthday.R
@@ -1,0 +1,5 @@
+#' @maestroFrequency 1 month
+#' @maestroStartTime 15 04:00:00
+my_monthly_pipeline <- function() {
+  TRUE
+}